rename service and add todos
(cherry picked from commit 6b72ba9692239e58aa7d5b6acf67f210eb702c06)
