+++ self_description_service.datasource/pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>datasource</artifactId>
+    <artifactId>self_description_service.datasource</artifactId>
     <version>${revision}</version>
     <packaging>jar</packaging>
     <build>
