+++ self_description_service.service/self_description_service.service.rest/pom.xml
@@ -137,7 +137,7 @@
         </dependency>
         <dependency>
             <groupId>org.eclipse.slm</groupId>
-            <artifactId>datasource</artifactId>
+            <artifactId>self_description_service.datasource</artifactId>
             <version>0.0.1-SNAPSHOT</version>
             <scope>compile</scope>
         </dependency>
